Move FWLiteAnalyzerWrapper.h to PhysicsTools/FWLite
The breaks the dependency of PhysicsTools/UtilAlgos on all FWLite
packages.
